test version

+++ defs.h
@@ -21,7 +21,7 @@ int decode_switches ( int argc, char * argv[], TSwitch * sw );
 void usage ( void );
 int read ( string filename, char mod );
 void match ( vector < unsigned int > * Occ );
-void Index ( unsigned int q, unsigned int size, unsigned int * M );
+void Index ( unsigned int q, long int size, unsigned int * M );
 string Num2Str ( unsigned int id, unsigned int length );
 unsigned int Str2Num ( string x );
 void factor ( int a, int q, string x, double p, vector<string> * list );
